requirements.txt
requests==2.27.1
mysql-connector-python==8.0.18
